requirements.txt
# Core dependencies
numpy>=1.24.3
pandas>=2.0.3
matplotlib>=3.7.2
seaborn>=0.12.2
scipy>=1.11.1

# Machine Learning
scikit-learn>=1.4.0     # scikit-learn==1.3.0 Scikit-learn 1.3.0 muss beim Installieren kompiliert werden und benötigt dafür einen C++ Compiler.
tensorflow>=2.16.0,<2.19.0

# Data processing
pyserial==3.5  # For Arduino serial communication

# Optional for advanced features
jupyter==1.0.0
plotly==5.15.0

# For NGIMU communication
python-osc==1.8.3
